set subdir variable to "po" in all po-*/Makefile.in.in files
This doesn't really make sense to me as that causes to always stamp the same file (po/stamp-it) instead of each po directory, but this is how intltoolize wants it.
At the same time this finally fixes the endless build loop if the source files are laying on xfs or tmpfs (which support more precise timestamps). (closes: #156)