Implementation of notification popups. (closes: #176)
Complete implementation of notification popups with action, icon, update,
append and sound support and as defined in
Desktop Notifications Specification 0.9 / 1.1:
http://www.galago-project.org/specs/notification/0.9/
http://people.canonical.com/~agateau/notifications-1.1/spec/
and Ubuntu's Notify-OSD Guidelines:
https://wiki.ubuntu.com/NotificationDevelopmentGuidelines

Tested successfully against notify-osd 0.9.23, notification-daemon 0.5.0, and
xfce4-notifyd 0.1.0.

Includes a hack for notification-daemon-xfce as that is not conformat with
the 0.9 specification and sends broken NotificationClosed dbus replies.