Fix openSUSE build.
The problem is that the openSUSE Nini package considers Nini.dll to be an
unstable assembly, and as per documented guidelines [0] it is not installed
into the GAC (which differs from Debian practice for this assembly).

Consequently, attempting to build Smuxi results in a compiler error, as
the NINI_LIBS value returned by pkg-config was '-r:Nini.dll', and Nini.dll
didn't exist within e.g. src/Engine/ at build time.

(An interesting side-note is that the openSUSE Build Service Smuxi package
explicitly copies Nini.dll into src/Engine, then sets MONO_PATH in order to
build Smuxi.  Using MONO_PATH is usually a hack, though, hence attempts to
"properly" fix it here.)

The workaround is to check the output of
`pkg-config --variable=Libraries nini-1.1`.  On Debian, this is the empty
string (as their nini-1.1.pc file doesn't have a Libraries variable), while on
openSUSE this is a file path.  If the length is non-zero, we assume openSUSE
semantics of a private assembly, and set NINI_LIBS accordingly.  Otherwise, we
continue with the existing Debian behavior of using NINI_LIBS as returned by
PKG_CHECK_MODULES() ~directly.

This also updates src/Engine/Makefile.am so that on openSUSE we deploy
Nini.dll into the installation directory (along with smuxi-engine.dll, etc.).

[0] http://www.mono-project.com/Guidelines:Application_Deployment