Default to using the _NetworkChat if the chat can't be found.
Allows bip to work because (for some bizarre reason) person messages are being
reported on the IRC info tab, e.g. the 'IRC balthasar-gimpnet' tab contains
messages such as:

	13:14 <kangaroo> 13:44:57]> right
	13:14 <kangaroo> 13:45:18]> brb, lunch

This makes no sense to me (they occurred on an actual channel, not...nothing),
but at least I can now connect with this fix.