Fixed connecting to an IRC server with SSL on a port other than 6697 via GUI (closes: #549)
It is not possible to connect to an IRC server using SSL unless it's on
port 6697.

This is because the port number is overwritten with 6697 whenever the
f_UseEncryptionCheckButton.Clicked event is called - and this event is called
whenever f_ValidateServerCertificateCheckButton is set to Active (GTK
ToggleButton Toggled signals are triggered by gtk_toggle_button_set_active).
So when the server configuration is loaded from the backend config file, the
port is set to the value in the file, then overwritten with 6697 if SSL is
enabled.

This patch fixes this, by setting the port number AFTER reading the SSL state
from the config file - so 6697 is set, then overwritten by the config file
value.