[Frontend-GNOME] Fixed last_focus_time window manager warnings (closes: #579)
<rant>
The API interaction of GTK+, X11 and freedesktop's notifications are driving
me nuts. The window manager will only move the focus of a window to another if
the timestamp of the event is new enough. And here begins the road of insanity.

The notification DBus API is not exposing the timestamp of the clicked action
button, neither is it possible to obtain the current timestamp of GTK as it only
exposes the timestamp of the currently _processed_ event via
gtk_get_current_event_time() [0] and guess what, as this is a D-Bus callback
gtk_get_current_event() is NULL and thus gtk_get_current_event_time() returns 0.
Of course the window focus will not change with a 0 timestamp. And no, you
can't just pass a unix timestamp here or something sane because then the
window manager will go crazy like this:

Window manager warning: last_focus_time (1300750862) is greater than comparison timestamp (4830419).  This most likely represents a buggy client sending inaccurate timestamps in messages such as _NET_ACTIVE_WINDOW.  Trying to work around...
Window manager warning: last_user_time (1300750862) is greater than comparison timestamp (4830419).  This most likely represents a buggy client sending inaccurate timestamps in messages such as _NET_ACTIVE_WINDOW.  Trying to work around...

After that it will stop to change the focus of any window for a while.... yay

The only way to "fix" this issue would be to fetch the current timestamp of X11
which is not possible with GTK+ because it only get's exposed by GDK's X11
specific API [1] which is not exposed at all in GTK#.

So we will have to live with this annoyance for now and not pass a
timestamp and use gtk_window_present() [2] instead.

Alternative would be to P/Invoke libgdk-x11-2.0.so directly which I am really
not in the mood for.
</rant>

[0] http://library.gnome.org/devel/gtk/stable/GtkWindow.html#gtk-window-present-with-time
[1] http://library.gnome.org/devel/gdk/stable/gdk-X-Window-System-Interaction.html#gdk-x11-get-server-time
[2] http://library.gnome.org/devel/gtk/stable/GtkWindow.html#gtk-window-present