[Engine] Implemented delayed auto-commit and re-indexing in Db4oMessageBuffer
Db4oMessageBuffer needs to commit changes back to db4o for newly added
messages. Making a commit when the chat is being disposed is not good enough
as that puts the risk of losing lots of messages in the case of a crash and
also uses lots of memory as db4o can only free the resources when the data
was written to disk.

For performance reasons the commit is delayed by using a memory buffer that
holds the messages. When the size of the buffer is exceeded the content of
the buffer gets committed to db4o. All methods that make use of the
db4o index have to honor this as it will no longer be the only source of
messages. At the same time we will no longer need to rebuild the db4o index
everytime a message was added. This is also a new performance improvement.