[Engine] Rewrite of Db4oMessageBuffer's index generation and db4o access
Dropped the memory backed IndexBuffer and db4o backed Index and replaced it
with a List<T> based Index. This new index keeps a reference to all
deactivated MessageModel objects. When messages are read they will be
activated, copied and then deactivated right away. The idea is to not leave
activated messages in the index which prevents memory to be freed again.
We can't just deactivate them without a copy as all references to the
previously activated message object would suddenly throw
NullReferenceExceptions because the members became null.
By copying the db4o objects we allow the GC to kick in when the message is
no longer used (referenced) by anything else.

This new index also allows us to incrementally add new messages to it without
the need to rebuild it at all. This was eating of lots of CPU cycles as it had
to query and sort all messages again from db4o with each flush.
We also simply save this index to db4o and load from it again when the
database is accessed. This needs some extra diskspace but is nothing compared
to all stored messages.