[Frontend-GNOME] Don't close tab on 3rd mouse button to follow GNOME convention
GNOME applications [0] are not using the 3rd mouse button to close tabs as
it is common to scroll through the tabs using the mouse wheel. Accidently
activating the mouse wheel button (3rd button) is pretty easy and leads to
annoyance of the user as he leaves a channel or a private conversation and
loses the message history with it.

 [0]: gedit, gnome-terminal and nautilus for example