[Frontend/Frontend-GNOME] Implemented background and concurrent chat sync (references: #375)
Instead of adding and syncing each chat at a time in the GUI thread and thus
block the GUI from refreshing, we are now using a thread pool with 4 worker
threads which add and sync the chats in the background and only populating the
retreived data is done in the GUI thread. This makes the frontend connect
faster, feels much more responsive and even allows to use the synced chats
already while it is still syncing the remaining ones.

Implemented the ChatViewSyncManager class which queues and processes the
AddChat and SyncChat calls using the ThreadPoolQueue.

Added Sync() to IChatView which the ChatViewSyncManager will call from a
thread pool worker thread and thus must be implemented thread-safely.

Added Populate() to IChatView which the ChatViewManager will call from it's
ChatViewSyncManager.ChatSynced handler to populate the data into the GUI.

Added ID and Position property to IChatView which the ChatViewManager will
call to populate the data before ChatViewSyncManager.ChatAdded will be raised
from a thread pool worker thread and thus must be implemented thread-safely.

Summary of the sync API workflow:

[Engine]
- FrontendManager.Sync() calls IFrontendUI.AddChat() and SyncChat() via
  .NET remoting.
[Frontend application]
- IFrontendUI.AddChat() calls ChatViewSyncManager.QueueAdd().
- IFrontendUI.SyncChat() calls ChatViewSyncManager.QueueSync().
[Frontend library]
- ChatViewSyncManager queues the Add and Sync requests to ThreadPoolQueue.
- ChatViewSyncManager.QueueAdd() create a sync lock to make sure a sync
  processor will not run before the chat was added to the frontend.
- ChatViewSyncManager's add worker will fetch ID, ChatType, Position and
  ProtocolManagerType from a thread pool worker thread and raises the
  ChatAdded event.
[Frontend application]
- ChatViewManager's ChatViewSyncManager.ChatAdded handler adds the chat to
  the GUI and calls ChatViewSyncManager.ReleaseSync() to signal the possibly
  waiting thread pool worker thread.
[Frontend library]
- ChatViewSyncManager's sync worker optionally honors a sync lock and then
  calls IChatView.Sync() from a thread pool worker thread and raises the
  ChatSynced event.
[Frontend application]
- IChatView.Sync() fetches all the data it needs from the ChatModel via
  .NET remoting.
- ChatViewManager's ChatViewSyncManager.ChatSynced handler calls
  IChatView.Populate() from the GUI thread.
- IChatView.Populate() pushes the fetched data into the GUI.