[Frontend-GNOME] Prepend GTK# to PATH on Windows (closes: #551)
GTK# on Windows relies to load the GTK+ libraries from the right location via
the PATH which it extends during the install. This works just fine as long as
PATH is not including any other GTK+ library paths. For this reason we simply
prepend the GTK#'s library path to the PATH environment variable when running
on Windows.