[Build] Detect target runtime/profile of compiler
On Ubuntu 11.10 with Mono 2.10 we have a xbuild that defaults to dmcs (4.0)
but Smuxi's configure script defaults to gmcs and this will lead to a mixed
2.0 and 4.0 build which is likely to crash the compiler.

We avoid this issue by detecting which target runtime/profile will be used and
then hint xbuild. This way make and xbuild will build for the same runtime.