[Frontend-GNOME] Disable forced NotificationAreaIconMode.Closed mode for Ubuntu's Messaging Menu
As of Ubuntu 12.10's libmessaging-menu, Smuxi can no longer restore a
hidden (closed) main window from messaging menu, as it relies on
single-instance behavior/support of the application [0]. The messaging menu
simply launches the application desktop file and expects that the application
detects the already running instance and restores a hidden/closed main window.

As of now Smuxi does not support single-instance as IPC on Mono and Linux
simply sucks.

Here the path that leads to clone wars:
 * You start Smuxi from launcher
 * You close the Smuxi window with the "X" button -> launcher no longer believes Smuxi is still running
 * libmessaging-menu no longer allows to bring Smuxi back (as it did with libindicate)
 * launcher and messaging menu will both now start new instances when you activate Smuxi
 * You close Smuxi's main window again with the "X" button
 * You activate Smuxi from the messaging menu or launcher -> you spawn another clone

 [0]: https://bugs.launchpad.net/indicator-messages/+bug/1065732/comments/1