[Frontend-GNOME] Support replying to persons by clicking on their nickname in the message area (closes: #771)
There are nicknames that are either difficult to remember and/or to type
(e.g. me4oslav). For this reason the user is now able to click on a nickname
in the message area and Smuxi automatically types that nick into the entry box.

Thanks goes to Marcelo Hashimoto (@conscioususer) for this nice idea!