[Engine-Campfire] Better handling for unkown messages
The JSON parser will give us a zeroed enum when it encounters a
message type that's not specified in the code. Change MessageType's
first value to be UnkownMessage instead of EnterMessage so we don't
show spurious join messages and instead show a message telling the
user that we received a message we don't understand.