[Engine(|-Test)] Improve URLs-in-brackets detection (closes #853).
URLs may now be prefixed by a specific set of delimiters (the common types
of brackets and quote marks), which are detected by the regular expression
but not considered part of the URL (thanks to match groups). Also, no
spurious space is added anymore if a URL is at the end of a line.