[Frontend-GNOME] Initial compile-only GTK#3 port + fully ported tool and join bar
All fields are that usually generated/provided by Stetic were created as stubs
to please the compiler. To avoid NullReferenceExceptions at runtime the classes
are guarded with a NotImplementedException thrown in the constructor.

Glade related code/types were disable with GLADE_SHARP define checks.

The tool bar of the MenuWidget was ported from Stetic to Glade3/GtkBuilder and
uses provides the same API as it was provided by Stetic. This avoid useless code
refactoring.
The menu bar of the MenuWidget was partially ported and needs more UI work with
Glade3.

The JoinWidget was also ported from Stetic to Glade3/GtkBuilder using the same
approach of the MenuWdiget, but is not showing up in the UI for some reason
(possibly a container or Show()-madness issue).