[.gitignore] Delete unnecessary nested .gitignore files
The .gitignore file in the root already contains the entry "bin/", so
it is not necessary to have these additional .gitignore files.

Their existence is a bit annoying as well because if you remove the
bin folder altogether (to make sure you have a clean checkout without
artifacts), `git status` would mark them as deleted.