Frontend-GNOME-XMPP: MessageType.Presence* didn't mark chatnodes with event color
HasEvent is only set for the contactlist and groupchats. Private chats are ignored on purpose:
In XmppPersonChatView only the most recent Presence state is shown (or for PresenceOnline nothing is shown). So it could have happened that a contact goes offline and back online, causing HasEvent to be set, but the MessageTextView of the XmppPersonChatView hasn't changed. Therefor the user would not see what has happened and wonder why his chatnode is colored in the event color.
contactlist and groupchats do not handle Presences in a special way, therefor these messages are always visible.