Engine: use compiled regexes in MessageBuilderSettings
This is not making a change with Mono (as of Mono up to 3.2.5) because the regex
compiler is disabled except MONO_NEW_RX=1 [0] is set as an environment variable.
Enabling this reduces the regex performance on Mono though. The regex compiler
is disabled by default because of uncertain regressions as noted here [1]. On
MS .NET this change should give a nice performance boost as seen here [2].

 [0]: https://github.com/mono/mono/blob/4db812ca7e6c4d4f8a43e6a40adeb8c111f0ddc0/mcs/class/System/System.Text.RegularExpressions/Regex.cs#L339
 [1]: https://bugzilla.novell.com/show_bug.cgi?id=470827
 [2]: http://stackoverflow.com/a/7707369