Engine(-IRC), Frontend-GNOME: support CertFP (closes: #96)
[CertFP][] is a NickServ authentication feature supported by modern IRC networks
as an more secure alternative to the famous "/msg NickServ IDENTIFY my_password"
command.

 [CertFP]: https://freenode.net/certfp/

As this is an internal setting only (for now) you need to configure it using
the /config command like this:

    /config Servers/IRC/$SERVER_ID/ClientCertificateFilename = mycert.pfx
    /config save

The client certificate can be generated using makecert like this:

    makecert -eku 1.3.6.1.5.5.7.3.2 -r -cy end -n "CN=$USER" -p12 mycert.pfx ""

The certificate must not use a passphrase, else it can't be loaded. Thus secure
the file against access by other users with:

    chmod 400 mycert.pfx

Place the certificate in ~/.config/smuxi/certs/ otherwise specify the full path
in ClientCertificateFilename.

On most IRC networks that support CertFP you can verify if the certificate was
used using /whois on your own nickname. A line like this should show up in the
whois reply:

    [276 (?) meebey3] has client certificate fingerprint a15aecab43e1d0965a2da43739a9628d790994e0

Special thanks goes to An-Ivoz for finding out how client certificate selection
works!