Common(-Tests): implemented SingleApplicationInstance
This implementation covers MS .NET on Windows, Mono with SHM on Unix, Mono
without SHM on Unix. It uses named mutex + Remoting or file mutex + Remoting
depending on what is supported by the operating system and CLI runtime.

 * MS .NET on Windows -> named mutex
 * Mono on Windows -> named mutex
 * Mono with SHM on Linux -> named mutex
 * Mono without SHM on Linux -> file lock

For file mutex support you need to compile this class with -define:MONO_UNIX
and -r:Mono.Posix.dll otherwise it will throw a NotSupportedException when
named mutex is not available.