[Common] IconCache: add method to download a particular image
Instead of just downloading a favicon, download a particular image from
a url directly. This will be useful for emoji and arbitrary external
images.

As this is the last part of what the favicon downloader does, we can
extract that functionality as well and use it for the direct downloader.