[Frontend-GNOME] Apply text size changes to emoji
When we resize the text, we must also take care to resize the pixbuf
containing the emoji, as those do not automatically become scaled.

To do this we must go through all emoji tags and replace their content
with a new pixbuf of the appropriate size. We must take care to re-tag
the new pixbuf with the tag so we can react to multiple size change
requests.