build: set libray path for 'run' target on OS X.
When we're on OS X, some of the libraries we need aren't readily
available from the system. We must therefore set up a fallback path for
the linker to look in mono's library dir, where e.g. libintl lives.

We do this in the script for Smuxi.app already, so let's use the same
paths in the Makefile.