Engine, Frontend*: introduced engine protocol version
Do not couple the release/assembly version with the protocol version as we can't
bump the Smuxi version 1.0 without this. All frontends would think the protocol
is no longer compatible because the major component of the assembly version does
not match with the frontend assembly version (0.x). Also refactored all code
that queried the now obsolete EngineVersion but still fallbacks to that for
backwards compatibility in case the smuxi-server is older than protocol 0.13.