Frontend-GNOME: mark messages of current chat as seen on quit (refs: #1058)
The issue is that Smuxi only marks messages as seen when the user switches away
from the current chat. This does not happen if the user was following the
currently visible chat and then quits Smuxi. Thus Smuxi has to mark the messages
of the current chat as seen on quit.