Server, Frontend-GNOME: use Boehm as GC instead of SGen (closes: #1062)
Mono 3.2.8 (and possibly newer versions) SEGVs in Mono.Data.Sqlite when the GC
is SGen and huge datasets are read from more than one thread at the same time
using different databases. With Boehm this does not happen. Since Smuxi switched
to SQLite by default this is a show stopper for the 1.0 release and thus Boehm
will be used to workaround this issue till the Mono SGen GC or the
Mono.Data.Sqlite binding will be fixed in a later Mono version. Or the used
SQLite binding will be replaced with sqlite-net [0] or Hyena's SQLite binding [1].

 [0]: https://github.com/praeclarum/sqlite-net
 [1]: https://github.com/GNOME/hyena/tree/master/Hyena.Data.Sqlite