Engine-Campfire: split messages on CR instead of trying to parse
We currently count the number of braces in order to determine when an
object definition is done. However, this has all sorts of problems
around quoting and strings.

It turns out that the server will send a CR as part of the newline it
sends between each message, which it won't produce inside the objects,
so we can split the message there and be confident that we're passing a
full object to the deserializer.