Engine: do not expose the user's realname automatically (refs: #541)
This is only a short term workaround. The user should have the option to use
their realname obtained from the operating system, but only after reviewing and
acknowledging that setting. This can be done using a dialog for the first start
which #541 wants to do.

This user information leak was also filed as [#779803][] in the Debian BTS.

 [#779803]: https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=779803