Engine(,Twitter): add ID to MessageModel (so that tweets have links)
In some engines such as Twitter, each message can contain a publicly
accessible URL associated to it. So we add the possibility to include
the ID of the tweet in MessageModel as a property, so that the
Frontend-GNOME-Twitter can form the URL needed to link from its
timestamp (inspired in how the Twitter web UI works: if you click on
the date of the message, you go to the URL of the single tweet).