configure.ac: check if shell supports +=
On systems with non-bash /bin/sh appending variables using += will fail. The
issue is that the configure script continues if it doesn't. This led to a
failing build in some obscure place because the variable wasn't appended. So we
bail now in the configure script early and give the user a hint how to fix it.