Common: copy win32 libraries to build directory
The native libraries in lib/win32/ are needed after the build else it will try
to load the libraries from the system and the versions probably don't match.

This actually happened with the SQLite library from Mono 4.x leading to an
obscure error like this:

    Exception Type:
    Mono.Data.Sqlite.SqliteException

    Exception Message:
    File opened that is not a database file
    file is encrypted or is not a database

    Exception StackTrace:
      at Mono.Data.Sqlite.SQLite3.Prepare (Mono.Data.Sqlite.SqliteConnection cnn, System.String strSql, Mono.Data.Sqlite.SqliteStatement previous, UInt32 timeoutMS, System.String& strRemain) [0x0022a] in C:\buildroot\release\repos\mono\mcs\class\Mono.Data.Sqlite\Mono.Data.Sqlite_2.0\SQLite3.cs:343
      at Mono.Data.Sqlite.SqliteCommand.BuildNextCommand () [0x00019] in C:\buildroot\release\repos\mono\mcs\class\Mono.Data.Sqlite\Mono.Data.Sqlite_2.0\SQLiteCommand.cs:230

The database wasn't damaged or anything but an incompatible SQLite library from
the system was loaded.