Frontend: rename method in ChatViewSyncManager, to follow convention
Methods that start with "On" are normally the ones that get
called because they got subscribed to some event. In this
case, this method is the one in charge of raising that event
so to make it a bit less confusing, let's rename it to have
the "Raise" prefix instead of "On".