Engine(-Twitter),Frontend: avoid Close warning dialog (closes: #1093)
There are some protocols, like Twitter, in which there are no
consequences from disconnecting from the server: no unavailable
status is shown to our contacts, and no chat buffers will be
lost because all the info resides server-side. For these, there
is not really any point in warning the user about current open
chats when she intends to close the connection, so let's
whitelist Twitter in this way to avoid the pointless dialog.