Engine(-IRC): prevent highlights from self (closes: #1095)
In theory, IRC is designed in a way in which all messages from
your current handle are always sent from your own client.

However, with the advent of IRC bridges, such as the Slack one,
it's feasible to use your IRC handle from 2 places: smuxi and
the web (slack UI). In this case, when you sent a message using
the latter, you would get a smuxi highlight on the former
because the engine receives a message that contains your nick,
which is essentially a false positive.

This can be workarounded by adding a comparison to Me's property
in MessageBuilder class just before highlighting.