Engine-Tests: mark failing NIE tests as Ignored("N/A")
Instead of having the build failing because the
tests throw NotImplementedException, we mark them
as Ignored for the subclasses where they don't
apply. The methods were not implemented because
the interface used for the MessageBuffer chosen
some time ago was IList, but it's too feature-rich
for these backends which, if implemented for them,
would be very expensive operations.