Core/Engines: migrate to MonoAddins infrastructure
Only for the backend for now.
(TODO: Load frontends with mono-addins too.)
This integration work opens the door for more
add-in management love, such as loading/unloading
capabilities at runtime, installing/updating
add-ins from remote repositories (via Cydin[1]), etc.

(The way to make this build in macOS is similar to
the approach done with log4net[2].)

[1] https://github.com/slluis/cydin
[2] bdbfb267d4bf66a8d0e59d0d85630379524797e6