Work around BSD install not having -D
The `-D` option is a GNU extension, so trying to build smuxi without GNU coreutils
is not going to work.

Since there's no indication from `./configure` that we need GNU install (as the
first thing it does is check for a BSD-compatible install), let's let it build
with the system's own userland.