Engine-MessageBuffer: support multiple source files for cat and copy actions
If someone wants to migrate existing message buffer files from multiple machines
that were running local Smuxi engines to a single smuxi-server one, it can be
useful to merge multiple message buffer files into a new one. The easiest way to
achieve this is by allowing to read from multiple buffer files in the same run
and simply concatenate all messages to the JSON pipe or a new destination
message buffer.

The use of IEnumerable<T>.Concat() can be pretty harmful but not in this
particular case as the stack height / recursion depth is only as height as the
number of message buffers were passed from the command line. For more details
why this can be dangerous, [see here][harmful-concat].

 [harmful-concat]: http://programmaticallyspeaking.com/how-enumerableconcat-brought-down-a-production-server.html