Server, Frontend-GNOME: fix Mono version detection for GC selection
The smuxi-server and smuxi-frontend-gnome start scripts incorrectly detect if
Mono is version 4 or higher. The grep call uses a hardcoded version 4 check
and thus Mono 5.x gets treated as it was Mono 3.x and thus it uses Boehm as GC
instead of SGen. Since Boehm is no longer really maintained/supported this can
lead easily to SEGVs on newer Mono versions. Now Boehm gets only enforced as GC
if Mono is 3.x or lower == SGen gets used if Mono is 4.x or newer.

Thanks goes to Llewelyn for reporting the issue on Mono 5.x