Engine(-Tests): support ignoring leading characters in message patterns
A message pattern could try to match a specific pattern like a word but wants to
remove/ignore leading characters. To remove and ignore trailing characters from
the match was already possible by using the DelimiterForEndOfPattern capture
group but not for leading characters. This feature is now extended by using a
capture group that is named DelimiterForStartOfPattern.

So for example a Smuxi message pattern with a regex of:

    (?<DelimiterForStartOfPattern>\s|\W)([a-fA-F0-9]{64})(?<DelimiterForEndOfPattern>\s|\W)

This will match bitcoin transaction hashes with a length of 64 hex numbers but
the whitespace (\s) and non-word characters (\W) in front or behind that won't
be part of linked result. This avoids false-positives when the pattern is likely
to be found within words.

There was also a bug in the existing DelimiterForEndOfPattern implementation
that would remove text from the {0} capture group which should contain the
complete text of the matched part. This is now fixed by retaining {0} to contain
the originally matched value and by using an explicit capture group {1} in the
heuristic domain regex.