Frontend-GNOME: handle unsuppported protocols nicely in ServerWidget
When older protocols get removed from Smuxi but they are still stored in the
config and thus available in the UI. The frontend should not throw an
ArgumentOutOfRangeException as that will lead to a forced shutdown of the
frontend after the user has closed the crash dialog. Instead just show that
the protocol is unsupported without any exceptions.