Engine: allow enabling emojis on-demand via MessageBuilderSettings.set_Emojis()
The old implementation only allowed to enable emojijs when the global setting
did so. Setting MessageBuilderSettings.Emojis to true after ApplyConfig() was
called had no effect as the property was just a flag. Now the Emojis property is
actually injecting the needed message pattern or removing it from the active
list of patterns.