Engine-IRC, Frontend-GNOME: add enhanced Slack support
Using the same approach as with the FacebookProtocolManager of having a
specialized protocol manager that is based on a generic one but with tweaks
specific to a service, like Facebook or Slack. This change introduces a
SlackProtocolManager that is based on the IrcProtocolManager.

Enable rendering emojis by default on Slack connections as emojis are natively
supported on Slack. (closes: #1103)