Frontend-GNOME: fix Gtk-CRITICAL messages when quitting
When quitting Smuxi, the terminal that runs the frontend will show Gtk-CRITICAL
messages like this:

    2017-06-11 14:31:35,035 [Main] DEBUG TRACE - [smuxi-frontend-gnome.exe] ChatViewManager.Clear()
    (smuxi-frontend-gnome:13501): Gtk-CRITICAL **: gtk_tree_store_get_path: assertion 'iter->user_data != NULL' failed
    (smuxi-frontend-gnome:13501): Gtk-CRITICAL **: IA__gtk_tree_model_get_iter: assertion 'path != NULL' failed
    2017-06-11 14:31:35,037 [Main] DEBUG TRACE - [smuxi-frontend-gnome.exe] ChatViewManager.OnTreeViewSelectionChanged(sender = Gtk.TreeSelection, e = System.EventArgs)

This happens because the ChatTreeView.Selection.Changed event handler assumes
there is always at least one item in the TreeView/Model which is not the case
during the shutdown which is clearing all items. So the last removed chat then
triggers this condition.