Engine-Tests: support assert fail messages for MessageBuilderTests.TestMessage()
When MessageBuilderTests.TestMessage() is called from a [Test] method in a
dynamic way, like inside a loop then you can't tell on which case / loop
iteration it failed for. Thus we support now an optional parameter for the
TestMessage() method that allows you to specify an assert fail message that is
used as a prefix for all the Assert.*() calls it performs.