Engine, Frontend-GNOME: use single remoting call for batched config changes
The preference dialog is pushing each config key and value to the UserConfig
class which then pushes each key to the remote engine. This was making the
preference dialog not responding for a long time with remote engines.

Now the preference dialog will collect all settings from the UI and then makes a
single remoting call via the newly added UserConfig.SetAll() method. Since this
is a new method exposed by the engine, the EngineProtocolVersion was bumped to
0.14. With older remote engines the preference dialog still needs to push each
key separately.