build: add push-transifex Makefile target
Also rename the previously committed `update-transifex`
to `pull-transifex` to distinguish better and sync with
the tx commands.

I just tested with Mirco that pushing an incomplete
translation to Transifex to a completed translation in
the server does not overwrite the work but it merges
it on the server, so it's fine if no -l parameter is
used here (meaning it defaults to pushing all langs).

--skip is needed because some translations may be
unfinished containing translation files that don't
have any translated strings, and the tx push command
would fail on those with the error "Could not import
file: We're not able to extract any string from the
file uploaded for language Persian (fa) in resource
Smuxi - IRC client: smuxi-server.pot"