Travis-CI: use official C# support with multiple Mono versions
Switch to the official [C# support of Travis][0]. When Smuxi started to use
Travis there was no native support for C# yet thus the C support was (ab)used
instead. The C build environment wasn't containing the Monon SDK of course so
that was done manually as part of the build steps.

 [0]: https://docs.travis-ci.com/user/languages/csharp/