Revert "Frontend-GNOME: restore cursor position in MessageTextView"
This reverts commit 22be95388e0aaae195538a62c8637f91db888692.

Turns out this change wasn't useful for blind persons as they rely on the
position of the cursor at all times. The red marker line as visual indicator
for marking were read messages end and new messages start is not useful to them
as the red marker line "jumps" from the current/last position to the bottom of
the chat if you leave the chat. This logic makes sense because the human eye can
easily jump many messages and still knows what is going on. For blind persons
they have to follow and go through each message one by one, otherwise they will
be lost of what happened in between. The cursor gives them the full control over
where they are and where they want to move next to, while an automatic jump
is completely confusing and useless to them.

The idea here was to align the "already read messages" logic from the red marker
line with the cursor position but instead it became more confusing and harder
for blind people. Thus this change is reverted and the exact position of the
cursor needs to be retained even across restarts of Smuxi, to be an improvement
for blind users.