Frontend-GNOME: workaround startup crash on Mac OS X with newer Mono versions
Smuxi no longer starts on OS X with Mono 5.10. Smuxi relies on the
libigemacintegration.dylib library to integrate into the Mac menu. This library
is shipped with the Mono runtime installation. Since Mono 5.10 this library no
longer provides all symbols that Smuxi needs and thus crashes like this:

	Exception Type:
	System.EntryPointNotFoundException

	Exception Message:
	ige_mac_menu_set_global_key_handler_enabled

	Exception StackTrace:
	  at (wrapper managed-to-native) IgeMacIntegration.IgeMacMenu.ige_mac_menu_set_global_key_handler_enabled(bool)
	  at IgeMacIntegration.IgeMacMenu.set_GlobalKeyHandlerEnabled (System.Boolean value) [0x00001] in <759e3226a68c4d99981055802cb0cde6>:0
	  at Smuxi.Frontend.Gnome.MenuWidget..ctor (Gtk.Window parent, Smuxi.Frontend.Gnome.ChatViewManager chatViewManager) [0x00261] in <759e3226a68c4d99981055802cb0cde6>:0
	  at Smuxi.Frontend.Gnome.MainWindow..ctor () [0x003a8] in <759e3226a68c4d99981055802cb0cde6>:0
	  at Smuxi.Frontend.Gnome.Frontend.Init (System.String[] args, System.String engine) [0x000cc] in <759e3226a68c4d99981055802cb0cde6>:0
	  at Smuxi.Frontend.Gnome.MainClass.Main (System.String[] args) [0x00235] in <759e3226a68c4d99981055802cb0cde6>:0

Really fixing this issue correctly is not straight forward because Smuxi should
bundle the library itself. So the workaround for now is to initialize the mac
menu integration with handling the exception and let Smuxi continue to start
without the mac menu integration enabled.

Thanks goes to Andres Aragoneses for reporting the issue, investigation and
testing the workaround.