Frontend-GNOME: fix open log buttons being sensitive after startup
When starting Smuxi the open log buttons are sensitive and can be activated.
Clicking one of the open log buttons directly after startup shows an error
dialog as there is no log file for the Smuxi session chat. The error message is:

    Cannot start process because a file name has not been provided.

The buttons should only be sensitive if the current active chat has a log file
though. Thus the initial state of the buttons will be set to being insensitive
and when switching to a different chat the buttons will be made sensitive in
MainWindow.OnNotebookSwitchPage().