Engine: fix serialization bug for CertFP support
The CertFP support in 83a2ab1c3 adds optional serialization and deserialization
of fields specific to the CertFP support. The _ClientCertificateFilename_ field
in particular was added to the serialization stream if the field is null or
empty.  Obviously it should only serialize it if the field is not null or empty.
The bug was never noticed and reported because there is no GUI to provide the
certificate file when doing an adhoc connection (instead of a saved one).