GitHub-CI/notify-irc: truncate commit message
Good commit messages are usually long with an explanation why the change was
made. To not spam the IRC channel with all the details, only send the first line
instead.

Unfortunately, GitHub Actions doesn't [functions][] to transform variables. Thus
we use a [forked version][] of the Action that does this transformation inside
the Python implementation for us.

  [functions]: https://docs.github.com/en/free-pro-team@latest/actions/reference/context-and-expression-syntax-for-github-actions#functions
  [forked version]: https://github.com/vmiklos/notify-irc