lib/Makefile.am: filter .git directories from source tarball
The `smuxi-1.1.tar.gz` tarballs contains a range of `.git` directories and files
that shouldn't be there as this is a tarball based source distribution.

This created a problem in Debian where `gbp import-orig smuxi-1.1.tar.gz` was
missing the files underneath `lib/SmartIrc4net` as it contained a full blown
git repo within the tarball.

The "dist-hook" automake target is used to filter such things from the resulting
source tarball that got included via the magic EXTRA_DIST variable.

The following `.git` files and directories are now removed from the tarball:

    ./lib/SmartIrc4net/.git
    ./lib/SmartIrc4net/src/starksoftproxy/.git
    ./lib/StarkSoftProxy/.git
    ./lib/agsxmpp/agsxmpp/starksoftproxy/.git