Build: fix Win32 builds with newer Mono versions
First symptom:

	Log Name:      Application
	Source:        .NET Runtime
	Date:          12/4/2023 10:32:39 AM
	Event ID:      1026
	Task Category: None
	Level:         Error
	Keywords:      Classic
	User:          N/A
	Computer:      DESKTOP-XXX
	Description:
	Application: smuxi-frontend-gnome.exe
	Framework Version: v4.0.30319
	Description: The process was terminated due to an unhandled exception.
	Exception Info: System.MissingMethodException
	   at NDesk.Options.Option..ctor(System.String, System.String, Int32)
	   at NDesk.Options.OptionSet.Add(System.String, System.String, System.Action`1<System.String>)
	   at Smuxi.Frontend.Gnome.MainClass.Main(System.String[])

This error made no sense, the methods are present (verified by
decompiling the binary) but the the MS .NET runtime insists the method
does not exist...

After commenting out the OptionSet.Add() calls, other methods could also
not be found:

	Exception Type:
	System.MissingMethodException

	Exception Message:
	Method not found: 'System.String[] System.String.Split(Char, System.StringSplitOptions)'.

	Exception StackTrace:
	   at Smuxi.Engine.Config.GetList(String key)
	   at Smuxi.Engine.Config.Get[T](String key, T defaultvalue)
	   at Smuxi.Engine.FrontendConfig.Load()
	   at Smuxi.Frontend.Gnome.Frontend.Init(String[] args, String engine)
	   at Smuxi.Frontend.Gnome.MainClass.Main(String[] args)

String.Split() is part of the core runtime library (mscorlib) but the 2nd
parameter StringSplitOptionswas only added in .NET standard 2.1 which
Smuxi never made use of! How could this happen? Newer Mono compiler version
is binding to a method that has parameter with a default value.

The actual C# codel is:

    result_str.Split('|');

But there is this definition in newer Mono/.NET versions:

    public string[] Split (char separator, StringSplitOptions options = System.StringSplitOptions.None);

So the C# compiler binds to this method signature instead:

    result_str.Split('|', System.StringSplitOptions.None);

Which then can't be found at runtime because smuxi-frontend-gnome.exe targets
the .NET 4.5 runtime which does not support .NET standard 2.1.

How can we prevent that the Mono C# compiler binds to .NET standard
methods breaking binary compatitibility with.NET 4.5?
Calling /usr/bin/dmcs as compiler (instead of /usr/bin/mcs) does the trick as
it enforces the .NET 4.0 SDK/runtime target. Thus we simply override the
MCS variable during configure in the dist-win32 make target.